Remove Vagrant management
I'm in the process of moving our VM bootstrap workflow over to Packer. If you
run Puppet within Packer and try to manage the Vagrant user, you'll get an
error from usermod: "usermod: user vagrant is currently logged in". There's
a script that sets this all up in the Packer workflow, so I'm removing the
classification in site.pp to defer to that script.